move RemoteTool implement to estellm package
