+++ spec_test.go
@@ -1,9 +1,11 @@
-package remote
+package estellm
 
 import (
 	"encoding/json"
 	"testing"
+	"time"
 
+	"github.com/Songmu/flextime"
 	"github.com/stretchr/testify/require"
 )
 
@@ -106,4 +108,36 @@ func TestSpecification_UnmarshalJSON_EmptyExtra(t *testing.T) {
 	require.JSONEq(t, `{"type": "object", "properties": {"input": {"type": "string"}}}`, string(spec.InputSchema))
 	require.Equal(t, "http://localhost:8080/worker", spec.WorkerEndpoint)
 	require.Empty(t, spec.Extra)
+}
+
+func TestSpecificationCache(t *testing.T) {
+	now := time.Now()
+	restore := flextime.Fix(now.AddDate(0, 0, -1))
+	defer restore()
+	cache := NewSpecificationCache(1 * time.Hour)
+
+	spec := Specification{Name: "test"}
+
+	// Set the specification in the cache
+	cache.Set("https://example.com/", spec)
+
+	// Get the specification from the cache
+	retrievedSpec, ok := cache.Get("https://example.com/")
+	require.True(t, ok)
+	require.Equal(t, spec, retrievedSpec)
+
+	flextime.Fix(now)
+
+	// Try to get the specification from the cache after expiration
+	_, ok = cache.Get("https://example.com/")
+	require.False(t, ok)
+	// Set the specification in the cache again
+	cache.Set("http://www.example.com/", spec)
+
+	// Delete the specification from the cache
+	cache.Delete("http://www.example.com/")
+
+	// Try to get the specification from the cache after deletion
+	_, ok = cache.Get("http://www.example.com/")
+	require.False(t, ok)
 }
